Fix cluster component cleanup for missing components, run cleanup before delete
The current cleanup threw a panic as the component object was nil.

We also have to run cleanup as part of the deletion process, to make sure we've
deleted any missing components first, as the process could stuck otherwise.
